DEV-1478: Restructure Row pre-populating page into 3 progressive examples
Split the single complex example into three progressive examples:
1. Basic spare rows (minSpareRows: 1 only)
2. Spare rows with placeholder styling (+ custom renderer)
3. Auto-populating with template values (existing full example, now example3)

All three examples use the same data for consistency.
