+++ docs/content/guides/rows/row-prepopulating/angular/example2.html
@@ -1,0 +1,3 @@
+<div>
+    <app-example2></app-example2>
+</div>
