+++ docs/content/guides/rows/row-prepopulating/angular/example3.html
@@ -1,0 +1,3 @@
+<div>
+    <app-example1></app-example1>
+</div>
